requirements.txt
# 基础依赖
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0

# 机器学习和深度学习
tensorflow>=2.9.0
scikit-learn>=1.0.0

# 数据可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 数据处理和分析
joblib>=1.1.0
python-dateutil>=2.8.2
tqdm>=4.64.0  # 进度条

# Web应用
Flask>=2.2.0

# 交易所API
python-binance>=1.0.16  # 币安API
ccxt>=2.0.0  # 多交易所支持

# 工具库
PyYAML>=6.0
requests>=2.28.0
schedule>=1.1.0

# 代理支持
PySocks>=1.7.1  # 用于SOCKS代理支持
aiohttp>=3.8.1  # 用于异步HTTP请求

# 数据库依赖
# SQLite (内置在Python中)
# MySQL/MariaDB
pymysql>=1.0.2
# PostgreSQL
psycopg2-binary>=2.9.3

# 警报和通知（可选）
twilio>=7.0.0  # 用于短信通知

# 技术指标（可选）
# ta-lib>=0.4.0  # 需要先安装系统依赖，如需使用请取消注释
